open feed type resolver for additional types
