+++ parser/feed/src/main/java/de/danoeh/antennapod/parser/feed/UnsupportedFeedtypeException.java
@@ -1,11 +1,11 @@
 package de.danoeh.antennapod.parser.feed;
 
-import de.danoeh.antennapod.parser.feed.util.TypeGetter;
-import de.danoeh.antennapod.parser.feed.util.TypeGetter.Type;
+import de.danoeh.antennapod.parser.feed.util.TypeResolver;
+import de.danoeh.antennapod.parser.feed.util.TypeResolver.Type;
 
 public class UnsupportedFeedtypeException extends Exception {
     private static final long serialVersionUID = 9105878964928170669L;
-    private final TypeGetter.Type type;
+    private final Type type;
     private String rootElement;
     private String message = null;
 
@@ -21,10 +21,10 @@ public class UnsupportedFeedtypeException extends Exception {
 
     public UnsupportedFeedtypeException(String message) {
         this.message = message;
-        type = Type.INVALID;
+        type = TypeResolver.Type.INVALID;
     }
 
-    public TypeGetter.Type getType() {
+    public Type getType() {
         return type;
     }
 
@@ -36,7 +36,7 @@ public class UnsupportedFeedtypeException extends Exception {
     public String getMessage() {
         if (message != null) {
             return message;
-        } else if (type == TypeGetter.Type.INVALID) {
+        } else if (type == TypeResolver.Type.INVALID) {
             return "Invalid type";
         } else {
             return "Type " + type + " not supported";
